Fixed 'not drawing a perfect circle' error when hissing.
As a side effect, it also fixed the 'not drawing in a straight line' error.  The issue was caused by me re-aligning the drawing mouse poisition with the position of the mouse pointer during every tick.  Since these two are apparently not quite the same coordinates, this caused the mouse to draw a spiral instead of a perfect circle when turning.